Update Plugins with Structure Info
Adding structure information (game folders) enables automatic detection of the main folder and simplifies installation. It also shows if a mod has folders that the game doesn't support. It unfortunately also marks mods as such, if they don't have any folders and only modify files in the main directory, but I don't think that's a problem.