Updated Tools Again
Changed wording, adding new discoveries about what could cause the error and how people might react to them. The detail about the Patch .zip in the plugin was left out, because it might cause people to extract this file inside MO2 and it's very unlikely that people have a plugin without the .zip but with the patch tool.